init: refactor auth to character
