+++ packages/backend/src/character/character.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { CharacterService } from './character.service';
+
+@Controller('character')
+export class CharacterController {
+  constructor(private readonly characterService: CharacterService) {}
+}
