+++ packages/backend/src/group/group.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { GroupService } from './group.service';
 import { GroupController } from './group.controller';
+import { CharacterRewardService } from '@/character/character-reward.service';
 
 @Module({
   controllers: [GroupController],
-  providers: [GroupService],
+  providers: [GroupService, CharacterRewardService],
 })
 export class GroupModule {}
